requirements.txt
flask
flask-cors
Flask-SQLAlchemy
PyJWT
Werkzeug
datasette
textblob
bcrypt
pandas